pyproject.toml
[project]
name = "sps-genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"   
dependencies = [
  "fastapi[standard]>=0.117.1",
  "pydantic>=2.11.9",
  "spacy>=3.8.7",
  "uvicorn>=0.36.0",
  "torch>=2.3,<2.6",
  "torchvision>=0.18,<0.21",
  "en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl",
  "matplotlib>=3.8",
  "Pillow>=10",
]
